+++ src/irmin-pack/IO.ml
@@ -18,12 +18,43 @@ let src = Logs.Src.create "irmin.pack.io" ~doc:"IO for irmin-pack"
 
 module Log = (val Logs.src_log src : Logs.LOG)
 
+let ( // ) = Filename.concat
+
+(* For every new version, update the [version] type and [versions]
+   headers. *)
+
+module Version = struct
+  type t = [ `V1 | `V2 ]
+
+  let enum = [ (`V1, "00000001"); (`V2, "00000002") ]
+
+  let pp = Fmt.of_to_string (function `V1 -> "v1" | `V2 -> "v2")
+
+  let to_bin v = List.assoc v enum
+
+  let raise_invalid v =
+    let pp_full_version ppf v = Fmt.pf ppf "%a (%S)" pp v (to_bin v) in
+    Fmt.failwith "invalid version: got %S, expecting %a" v
+      Fmt.(Dump.list pp_full_version)
+      (List.map fst enum)
+
+  let of_bin b =
+    try Some (List.assoc b (List.map (fun (x, y) -> (y, x)) enum))
+    with Not_found -> None
+end
+
+type version = Version.t
+
+let pp_version = Version.pp
+
+exception Invalid_version of { expected : version; found : version }
+
 module type S = sig
   type t
 
   exception RO_Not_Allowed
 
-  val v : fresh:bool -> version:string -> readonly:bool -> string -> t
+  val v : version:version option -> fresh:bool -> readonly:bool -> string -> t
 
   val name : t -> string
 
@@ -39,13 +70,23 @@ module type S = sig
 
   val force_offset : t -> int64
 
+  val generation : t -> int64
+
+  val force_generation : t -> int64
+
   val readonly : t -> bool
 
-  val version : t -> string
+  val version : t -> version
 
   val flush : t -> unit
 
   val close : t -> unit
+
+  val migrate :
+    progress:(int64 -> unit) ->
+    t ->
+    version ->
+    (unit, [> `Msg of string ]) result
 end
 
 let ( ++ ) = Int64.add
@@ -60,16 +101,19 @@ module Unix : S = struct
   type t = {
     file : string;
     mutable raw : Raw.t;
+    mutable generation : int64;
     mutable offset : int64;
     mutable flushed : int64;
     readonly : bool;
-    version : string;
+    version : version;
     buf : Buffer.t;
   }
 
   let name t = t.file
 
-  let header = 16L (* offset + version *)
+  let header = function
+    | `V1 -> (* offset + version *) 16L
+    | `V2 -> (* offset + version + generation *) 24L
 
   let unsafe_flush t =
     Log.debug (fun l -> l "IO flush %s" t.file);
@@ -82,11 +126,11 @@ module Unix : S = struct
       Raw.Offset.set t.raw offset;
       (* concurrent append might happen so here t.offset might differ
          from offset *)
-      if not (t.flushed ++ Int64.of_int (String.length buf) = header ++ offset)
-      then
+      let h = header t.version in
+      if not (t.flushed ++ Int64.of_int (String.length buf) = h ++ offset) then
         Fmt.failwith "sync error: %s flushed=%Ld offset+header=%Ld\n%!" t.file
-          t.flushed (offset ++ header);
-      t.flushed <- offset ++ header
+          t.flushed (offset ++ h);
+      t.flushed <- offset ++ h
 
   let flush t =
     if t.readonly then raise RO_Not_Allowed;
@@ -103,15 +147,16 @@ module Unix : S = struct
   let set t ~off buf =
     if t.readonly then raise RO_Not_Allowed;
     unsafe_flush t;
-    Raw.unsafe_write t.raw ~off:(header ++ off) buf;
+    Raw.unsafe_write t.raw ~off:(header t.version ++ off) buf;
     assert (
       let len = Int64.of_int (String.length buf) in
-      let off = header ++ off ++ len in
+      let off = header t.version ++ off ++ len in
       off <= t.flushed)
 
   let read t ~off buf =
-    assert (if not t.readonly then header ++ off <= t.flushed else true);
-    Raw.unsafe_read t.raw ~off:(header ++ off) ~len:(Bytes.length buf) buf
+    let off = header t.version ++ off in
+    assert (if not t.readonly then off <= t.flushed else true);
+    Raw.unsafe_read t.raw ~off ~len:(Bytes.length buf) buf
 
   let offset t = t.offset
 
@@ -119,7 +164,16 @@ module Unix : S = struct
     t.offset <- Raw.Offset.get t.raw;
     t.offset
 
-  let version t = t.version
+  let generation t = t.generation
+
+  let force_generation t =
+    t.generation <- Raw.Generation.get t.raw;
+    t.generation
+
+  let version t =
+    Log.debug (fun l ->
+        l "[%s] version: %a" (Filename.basename t.file) pp_version t.version);
+    t.version
 
   let readonly t = t.readonly
 
@@ -146,14 +200,54 @@ module Unix : S = struct
     in
     aux dirname (fun () -> ())
 
-  let clear t =
-    t.offset <- 0L;
-    t.flushed <- header;
-    Buffer.clear t.buf
+  let raw ~flags ~version ~offset ~generation file =
+    let x = Unix.openfile file flags 0o644 in
+    let raw = Raw.v x in
+    let header =
+      { Raw.Header.version = Version.to_bin version; offset; generation }
+    in
+    Raw.Header.set raw header;
+    raw
 
-  let v ~fresh ~version:current_version ~readonly file =
-    assert (String.length current_version = 8);
-    let v ~offset ~version raw =
+  let unsafe_clear t =
+    if t.readonly then invalid_arg "Read-only IO cannot be cleared";
+    Log.debug (fun l -> l "clear %s" t.file);
+    Buffer.clear t.buf;
+    (* no-op if the file is already empty; this is to avoid bumping
+       the version number when this is not necessary. *)
+    if t.offset = 0L then ()
+    else (
+      t.offset <- 0L;
+      t.generation <- Int64.succ t.generation;
+      t.flushed <- header t.version;
+      (* update the generation for concurrent readonly instance to
+         notice that the file has been clear when they next sync. *)
+      Raw.Generation.set t.raw t.generation;
+      (* delete the file. *)
+      Raw.close t.raw;
+      Unix.unlink t.file;
+      (* and re-open a fresh instance. *)
+      t.raw <-
+        raw ~version:t.version ~generation:t.generation ~offset:0L
+          ~flags:Unix.[ O_CREAT; O_RDWR; O_CLOEXEC ]
+          t.file)
+
+  let clear t =
+    match t.version with
+    | `V1 -> invalid_arg "V1 stores cannot be cleared"
+    | `V2 -> unsafe_clear t
+
+  let v ~version ~fresh ~readonly file =
+    let get_version () =
+      match version with
+      | Some v -> v
+      | None ->
+          Fmt.invalid_arg
+            "Must supply an explicit version when creating a new store ({ file \
+             = %s })"
+            file
+    in
+    let v ~offset ~version ~generation raw =
       {
         version;
         file;
@@ -161,64 +255,149 @@ module Unix : S = struct
         raw;
         readonly;
         buf = Buffer.create (4 * 1024);
-        flushed = header ++ offset;
+        flushed = header version ++ offset;
+        generation;
       }
     in
     let mode = Unix.(if readonly then O_RDONLY else O_RDWR) in
     mkdir (Filename.dirname file);
     match Sys.file_exists file with
     | false ->
-        let x = Unix.openfile file Unix.[ O_CREAT; mode; O_CLOEXEC ] 0o644 in
-        let raw = Raw.v x in
-        Raw.Offset.set raw 0L;
-        Raw.Version.set raw current_version;
-        v ~offset:0L ~version:current_version raw
-    | true ->
+        let version = get_version () in
+        let raw =
+          raw
+            ~flags:[ O_CREAT; mode; O_CLOEXEC ]
+            ~version ~offset:0L ~generation:0L file
+        in
+        v ~offset:0L ~version ~generation:0L raw
+    | true -> (
         let x = Unix.openfile file Unix.[ O_EXCL; mode; O_CLOEXEC ] 0o644 in
         let raw = Raw.v x in
         if fresh then (
-          Raw.Offset.set raw 0L;
-          Raw.Version.set raw current_version;
-          v ~offset:0L ~version:current_version raw)
+          let version = get_version () in
+          let header =
+            {
+              Raw.Header.version = Version.to_bin version;
+              offset = 0L;
+              generation = 0L;
+            }
+          in
+          Raw.Header.set raw header;
+          v ~offset:0L ~version ~generation:0L raw)
         else
-          let offset = Raw.Offset.get raw in
-          let version = Raw.Version.get raw in
-          assert (version = current_version);
-          v ~offset ~version raw
+          let actual_version =
+            let v_string = Raw.Version.get raw in
+            match Version.of_bin v_string with
+            | Some v -> v
+            | None -> Version.raise_invalid v_string
+          in
+          (match version with
+          | Some v when v <> actual_version ->
+              raise (Invalid_version { expected = v; found = actual_version })
+          | _ -> ());
+          match actual_version with
+          | `V1 ->
+              Log.debug (fun l -> l "[%s] file exists in V1" file);
+              let offset = Raw.Offset.get raw in
+              v ~offset ~version:`V1 ~generation:0L raw
+          | `V2 ->
+              let { Raw.Header.offset; generation; _ } = Raw.Header.get raw in
+              v ~offset ~version:`V2 ~generation raw)
 
   let close t = Raw.close t.raw
+
+  (* From a given offset in [src], transfer all data to [dst] (starting at
+     [dst_off]). *)
+  let transfer_all ~progress ~src ~src_off ~dst ~dst_off =
+    let ( + ) a b = Int64.(add a (of_int b)) in
+    let buf_len = 4096 in
+    let buf = Bytes.create buf_len in
+    let rec inner ~src_off ~dst_off =
+      match Raw.unsafe_read src ~off:src_off ~len:buf_len buf with
+      | 0 -> ()
+      | read ->
+          assert (read <= buf_len);
+          let to_write = if read < buf_len then Bytes.sub buf 0 read else buf in
+          let () =
+            Raw.unsafe_write dst ~off:dst_off (Bytes.unsafe_to_string to_write)
+          in
+          progress (Int64.of_int read);
+          (inner [@tailcall]) ~src_off:(src_off + read) ~dst_off:(dst_off + read)
+    in
+    inner ~src_off ~dst_off
+
+  let migrate ~progress src dst_v =
+    let src_v =
+      let v_bin = Raw.Version.get src.raw in
+      match Version.of_bin v_bin with
+      | None -> Fmt.failwith "Could not parse version string `%s'" v_bin
+      | Some v -> v
+    in
+    let src_offset = Raw.Offset.get src.raw in
+    match (src_v, dst_v) with
+    | `V1, `V2 ->
+        let dst_path =
+          let rand = Random.State.(bits (make_self_init ())) land 0xFFFFFF in
+          Fmt.str "%s-tmp-migrate-%06x" src.file rand
+        in
+        Log.debug (fun m ->
+            m "[%s] Performing migration [%a → %a] using tmp file %s"
+              (Filename.basename src.file)
+              pp_version `V1 pp_version `V2
+              (Filename.basename dst_path));
+        let dst =
+          (* Note: all V1 files implicitly have [generation = 0], since it
+             is not possible to [clear] them. *)
+          raw dst_path ~flags:[ Unix.O_CREAT; O_WRONLY ] ~version:`V2
+            ~offset:src_offset ~generation:0L
+        in
+        transfer_all ~src:src.raw ~progress
+          ~src_off:(header `V1)
+          ~dst
+          ~dst_off:(header `V2);
+        Raw.close dst;
+        Unix.rename dst_path src.file;
+        Ok ()
+    | _, _ ->
+        Fmt.invalid_arg "[%s] Unsupported migration path: %a → %a"
+          (Filename.basename src.file)
+          pp_version src_v pp_version dst_v
 end
 
-let ( // ) = Filename.concat
+module Cache = struct
+  type ('a, 'v) t = { v : 'a -> ?fresh:bool -> ?readonly:bool -> string -> 'v }
 
-let with_cache ~v ~clear ~valid file =
-  let files = Hashtbl.create 13 in
-  let cached_constructor extra_args ?(fresh = false) ?(readonly = false) root =
-    let file = root // file in
-    if fresh && readonly then invalid_arg "Read-only IO cannot be fresh";
-    try
-      if not (Sys.file_exists file) then (
+  let memoize ~v ~clear ~valid file =
+    let files = Hashtbl.create 13 in
+    let cached_constructor extra_args ?(fresh = false) ?(readonly = false) root
+        =
+      let file = root // file in
+      if fresh && readonly then invalid_arg "Read-only IO cannot be fresh";
+      try
+        if not (Sys.file_exists file) then (
+          Log.debug (fun l ->
+              l "[%s] does not exist anymore, cleaning up the fd cache"
+                (Filename.basename file));
+          Hashtbl.remove files (file, true);
+          Hashtbl.remove files (file, false);
+          raise Not_found);
+        let t = Hashtbl.find files (file, readonly) in
+        if valid t then (
+          Log.debug (fun l ->
+              l "Found in cache: { path = %s; readonly = %b }" file readonly);
+          if fresh then clear t;
+          t)
+        else (
+          Hashtbl.remove files (file, readonly);
+          raise Not_found)
+      with Not_found ->
         Log.debug (fun l ->
-            l "[%s] does not exist anymore, cleaning up the fd cache"
-              (Filename.basename file));
-        Hashtbl.remove files (file, true);
-        Hashtbl.remove files (file, false);
-        raise Not_found);
-      let t = Hashtbl.find files (file, readonly) in
-      if valid t then (
-        Log.debug (fun l -> l "%s found in cache" file);
+            l "[%s] v fresh=%b readonly=%b" (Filename.basename file) fresh
+              readonly);
+        let t = v extra_args ~fresh ~readonly file in
         if fresh then clear t;
-        t)
-      else (
-        Hashtbl.remove files (file, readonly);
-        raise Not_found)
-    with Not_found ->
-      Log.debug (fun l ->
-          l "[%s] v fresh=%b readonly=%b" (Filename.basename file) fresh
-            readonly);
-      let t = v extra_args ~fresh ~readonly file in
-      if fresh then clear t;
-      Hashtbl.add files (file, readonly) t;
-      t
-  in
-  `Staged cached_constructor
+        Hashtbl.add files (file, readonly) t;
+        t
+    in
+    { v = cached_constructor }
+end